backend/requirements.txt
fastapi
uvicorn[standard]
supabase
openai
python-dotenv
numpy
httpx
